move tests to the right folder and file
