+++ tracer/test/Datadog.Trace.ClrProfiler.Managed.Tests/AutoInstrumentation/AWS/SNS/MemoryStreamCarrierGetter.cs
